Conflicts:
	Projects/LevelPerformanceTest/Classes/GameCore.cpp

Former-commit-id: da8d08b4868ea72d820545a072abae39f8d1aa33

+++ Sources/Internal/Platform/Qt/MacOS/QtLayerMacOS.h
@@ -66,7 +66,7 @@ public:
     
     void InitializeGlWindow(void *qtView, int32 width, int32 height);
     
-    void MouseMoved(int32 x, int32 y);
+    void MouseMoved(float32 x, float32 y);
 	
 	virtual void* GetOpenGLView();
 };
